+++ demo-app/src/main/kotlin/io/getstream/video/android/data/datasource/local/InMemoryStore.kt
@@ -1,0 +1,10 @@
+package io.getstream.video.android.data.datasource.local
+
+import io.getstream.video.android.model.User
+
+interface InMemoryStore {
+
+    fun saveUser(user: User)
+
+    fun getUser(): User?
+}
