+++ demo-app/src/main/kotlin/io/getstream/video/android/data/datasource/local/NoopMemoryStoreImpl.kt
@@ -1,0 +1,10 @@
+package io.getstream.video.android.data.datasource.local
+
+import io.getstream.video.android.model.User
+
+class NoopMemoryStoreImpl : InMemoryStore {
+
+    override fun saveUser(user: User) {}
+
+    override fun getUser(): User? = null
+}
